Consolidate SDK location
