--- sdk/bondmcp_python/async_client.py
@@ -1,147 +1,0 @@
-"""
-BondMCP Python SDK Async Client
-
-Async client implementation for the BondMCP API.
-"""
-
-import asyncio
-
-class AsyncBondMCPClient:
-    """
-    Async client for the BondMCP API.
-    
-    This client provides asynchronous methods to interact with the BondMCP API.
-    """
-    
-    def __init__(self, api_key, base_url="https://api.bondmcp.com", timeout=30, max_retries=3, retry_delay=2, http_client=None):
-        """
-        Initialize a new async BondMCP client.
-        
-        Args:
-            api_key (str): Your BondMCP API key
-            base_url (str, optional): Base URL for the API. Defaults to "https://api.bondmcp.com".
-            timeout (int, optional): Request timeout in seconds. Defaults to 30.
-            max_retries (int, optional): Maximum number of retry attempts. Defaults to 3.
-            retry_delay (int, optional): Delay between retries in seconds. Defaults to 2.
-            http_client (object, optional): Custom HTTP client. Defaults to None.
-        """
-        self.api_key = api_key
-        self.base_url = base_url
-        self.timeout = timeout
-        self.max_retries = max_retries
-        self.retry_delay = retry_delay
-        self.http_client = http_client
-        
-        # Initialize sub-clients
-        self.labs = AsyncLabsClient(self)
-        self.health_data = AsyncHealthDataClient(self)
-        self.supplement = AsyncSupplementClient(self)
-        self.webhooks = AsyncWebhookClient(self)
-    
-    async def ask(self, message, context=None, include_sources=True, max_tokens=500, user_id=None):
-        """
-        Ask a health-related question asynchronously.
-        
-        Args:
-            message (str): The health question to ask
-            context (str, optional): Context for the question. Defaults to None.
-            include_sources (bool, optional): Whether to include reference sources. Defaults to True.
-            max_tokens (int, optional): Maximum response length. Defaults to 500.
-            user_id (str, optional): Optional user identifier for personalization. Defaults to None.
-            
-        Returns:
-            dict: The response containing the answer and metadata
-        """
-        # Implementation would go here
-        await asyncio.sleep(0.1)  # Simulate API call
-        return {"answer": "Sample response", "confidence": 0.95}
-    
-    async def close(self):
-        """Close the client and release resources."""
-        if hasattr(self.http_client, 'close'):
-            await self.http_client.close()
-
-
-class AsyncLabsClient:
-    """Async client for lab result interpretation endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    async def interpret(self, lab_results):
-        """
-        Interpret multiple lab results asynchronously.
-        
-        Args:
-            lab_results (list): List of lab result dictionaries
-            
-        Returns:
-            dict: Interpretation results
-        """
-        # Implementation would go here
-        await asyncio.sleep(0.1)  # Simulate API call
-        return {"insights": "Sample insights", "recommendations": []}
-
-
-class AsyncHealthDataClient:
-    """Async client for health data analysis endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    async def analyze(self, data):
-        """
-        Analyze health data asynchronously.
-        
-        Args:
-            data (dict): Health data to analyze
-            
-        Returns:
-            dict: Analysis results
-        """
-        # Implementation would go here
-        await asyncio.sleep(0.1)  # Simulate API call
-        return {"analysis": "Sample analysis", "recommendations": []}
-
-
-class AsyncSupplementClient:
-    """Async client for supplement recommendation endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    async def recommend(self, health_goals):
-        """
-        Get supplement recommendations based on health goals asynchronously.
-        
-        Args:
-            health_goals (list): List of health goals
-            
-        Returns:
-            dict: Supplement recommendations
-        """
-        # Implementation would go here
-        await asyncio.sleep(0.1)  # Simulate API call
-        return {"supplements": []}
-
-
-class AsyncWebhookClient:
-    """Async client for webhook management endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    async def create(self, url, events):
-        """
-        Register a webhook asynchronously.
-        
-        Args:
-            url (str): Webhook URL
-            events (list): List of events to subscribe to
-            
-        Returns:
-            dict: Webhook details
-        """
-        # Implementation would go here
-        await asyncio.sleep(0.1)  # Simulate API call
-        return {"id": "webhook_123", "url": url, "events": events}
