--- sdk/bondmcp_python/client.py
@@ -1,229 +1,0 @@
-"""
-BondMCP Python SDK Client
-
-Main client class for interacting with the BondMCP API.
-"""
-
-class BondMCPClient:
-    """
-    Client for the BondMCP API.
-    
-    This client provides methods to interact with the BondMCP API for healthcare AI applications.
-    """
-    
-    def __init__(self, api_key, base_url="https://api.bondmcp.com", timeout=30, max_retries=3, retry_delay=2, http_client=None):
-        """
-        Initialize a new BondMCP client.
-        
-        Args:
-            api_key (str): Your BondMCP API key
-            base_url (str, optional): Base URL for the API. Defaults to "https://api.bondmcp.com".
-            timeout (int, optional): Request timeout in seconds. Defaults to 30.
-            max_retries (int, optional): Maximum number of retry attempts. Defaults to 3.
-            retry_delay (int, optional): Delay between retries in seconds. Defaults to 2.
-            http_client (object, optional): Custom HTTP client. Defaults to None.
-        """
-        self.api_key = api_key
-        self.base_url = base_url
-        self.timeout = timeout
-        self.max_retries = max_retries
-        self.retry_delay = retry_delay
-        self.http_client = http_client
-        
-        # Initialize sub-clients
-        self.labs = LabsClient(self)
-        self.health_data = HealthDataClient(self)
-        self.supplement = SupplementClient(self)
-        self.webhooks = WebhookClient(self)
-    
-    def ask(self, message, context=None, include_sources=True, max_tokens=500, user_id=None):
-        """
-        Ask a health-related question.
-        
-        Args:
-            message (str): The health question to ask
-            context (str, optional): Context for the question. Defaults to None.
-            include_sources (bool, optional): Whether to include reference sources. Defaults to True.
-            max_tokens (int, optional): Maximum response length. Defaults to 500.
-            user_id (str, optional): Optional user identifier for personalization. Defaults to None.
-            
-        Returns:
-            dict: The response containing the answer and metadata
-        """
-        # Implementation would go here
-        return {"answer": "Sample response", "confidence": 0.95}
-
-
-class LabsClient:
-    """Client for lab result interpretation endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    def interpret(self, lab_results):
-        """
-        Interpret multiple lab results.
-        
-        Args:
-            lab_results (list): List of lab result dictionaries
-            
-        Returns:
-            dict: Interpretation results
-        """
-        # Implementation would go here
-        return {"insights": "Sample insights", "recommendations": []}
-    
-    def interpret_single(self, name, value, unit, reference_range=None):
-        """
-        Interpret a single lab result.
-        
-        Args:
-            name (str): Lab test name
-            value (float): Lab test value
-            unit (str): Unit of measurement
-            reference_range (str, optional): Reference range. Defaults to None.
-            
-        Returns:
-            dict: Interpretation results
-        """
-        # Implementation would go here
-        return {"insight": "Sample insight", "recommendation": ""}
-
-
-class HealthDataClient:
-    """Client for health data analysis endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    def analyze(self, data):
-        """
-        Analyze health data.
-        
-        Args:
-            data (dict): Health data to analyze
-            
-        Returns:
-            dict: Analysis results
-        """
-        # Implementation would go here
-        return {"analysis": "Sample analysis", "recommendations": []}
-    
-    def list(self, page=1, limit=10):
-        """
-        List health data entries.
-        
-        Args:
-            page (int, optional): Page number. Defaults to 1.
-            limit (int, optional): Items per page. Defaults to 10.
-            
-        Returns:
-            dict: Paginated list of entries
-        """
-        # Implementation would go here
-        return {"entries": [], "total": 0, "page": page, "limit": limit}
-    
-    def list_all(self):
-        """
-        Generator to iterate through all health data entries.
-        
-        Yields:
-            dict: Page of entries
-        """
-        page = 1
-        while True:
-            result = self.list(page=page)
-            yield result
-            if len(result["entries"]) < result["limit"]:
-                break
-            page += 1
-
-
-class SupplementClient:
-    """Client for supplement recommendation endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    def recommend(self, health_goals):
-        """
-        Get supplement recommendations based on health goals.
-        
-        Args:
-            health_goals (list): List of health goals
-            
-        Returns:
-            dict: Supplement recommendations
-        """
-        # Implementation would go here
-        return {"supplements": []}
-    
-    def recommend_for_labs(self, lab_results):
-        """
-        Get supplement recommendations based on lab results.
-        
-        Args:
-            lab_results (list): List of lab result dictionaries
-            
-        Returns:
-            dict: Supplement recommendations
-        """
-        # Implementation would go here
-        return {"supplements": []}
-    
-    def get_info(self, supplement_name):
-        """
-        Get detailed information about a specific supplement.
-        
-        Args:
-            supplement_name (str): Name of the supplement
-            
-        Returns:
-            dict: Supplement information
-        """
-        # Implementation would go here
-        return {"name": supplement_name, "description": "", "dosage": "", "warnings": []}
-
-
-class WebhookClient:
-    """Client for webhook management endpoints."""
-    
-    def __init__(self, parent_client):
-        self.parent_client = parent_client
-    
-    def create(self, url, events):
-        """
-        Register a webhook.
-        
-        Args:
-            url (str): Webhook URL
-            events (list): List of events to subscribe to
-            
-        Returns:
-            dict: Webhook details
-        """
-        # Implementation would go here
-        return {"id": "webhook_123", "url": url, "events": events}
-    
-    def list(self):
-        """
-        List registered webhooks.
-        
-        Returns:
-            list: List of webhooks
-        """
-        # Implementation would go here
-        return []
-    
-    def delete(self, webhook_id):
-        """
-        Delete a webhook.
-        
-        Args:
-            webhook_id (str): Webhook ID
-            
-        Returns:
-            bool: Success status
-        """
-        # Implementation would go here
-        return True
