--- sdk/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "bondmcp-js",
-  "version": "0.1.0",
-  "description": "Node.js SDK for the BondMCP API",
-  "main": "bondmcp-node.js",
-  "type": "commonjs",
-  "files": [
-    "bondmcp-node.js"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/bondmcp/mcp"
-  },
-  "keywords": [
-    "bondmcp",
-    "sdk",
-    "api"
-  ],
-  "author": "BondMCP",
-  "license": "MIT"
-}
