--- sdks/typescript/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "@bondmcp/typescript-sdk",
-  "version": "2.0.0",
-  "description": "Official TypeScript/JavaScript SDK for BondMCP - The #1 MCP in Health",
-  "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src",
-    "README.md",
-    "LICENSE"
-  ],
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "lint": "eslint src --ext .ts,.tsx",
-    "lint:fix": "eslint src --ext .ts,.tsx --fix",
-    "type-check": "tsc --noEmit",
-    "docs": "typedoc src/index.ts",
-    "prepublishOnly": "npm run build"
-  },
-  "keywords": [
-    "bondmcp",
-    "health",
-    "medical",
-    "ai",
-    "api",
-    "sdk",
-    "typescript",
-    "javascript",
-    "healthcare",
-    "mcp"
-  ],
-  "author": "BondMCP Team <support@bondmcp.com>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/bondmcp/mcp.git",
-    "directory": "sdks/typescript"
-  },
-  "homepage": "https://bondmcp.com",
-  "bugs": {
-    "url": "https://github.com/bondmcp/mcp/issues"
-  },
-  "dependencies": {
-    "axios": "^1.6.0",
-    "zod": "^3.22.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.0",
-    "@types/node": "^20.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "eslint": "^8.50.0",
-    "jest": "^29.7.0",
-    "rollup": "^4.0.0",
-    "@rollup/plugin-typescript": "^11.1.0",
-    "@rollup/plugin-node-resolve": "^15.2.0",
-    "@rollup/plugin-commonjs": "^25.0.0",
-    "rollup-plugin-dts": "^6.1.0",
-    "ts-jest": "^29.1.0",
-    "typedoc": "^0.25.0",
-    "typescript": "^5.2.0"
-  },
-  "engines": {
-    "node": ">=16.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
-
