+++ tests/test_cli.py
@@ -1,6 +1,11 @@
 from click.testing import CliRunner
 import sys
+from pathlib import Path
 import types
+
+ROOT = Path(__file__).resolve().parent.parent
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
 
 if "requests" not in sys.modules:
     requests_stub = types.ModuleType("requests")
