+++ tests/test_client.py
@@ -1,20 +1,14 @@
 import json
-import importlib.util
-from pathlib import Path
 import sys
+from pathlib import Path
 import requests
 import pytest
 
 ROOT = Path(__file__).resolve().parent.parent
-SDK_DIR = ROOT / "sdk"
-if str(SDK_DIR) not in sys.path:
-    sys.path.insert(0, str(SDK_DIR))
-spec = importlib.util.spec_from_file_location(
-    "sdk.bondmcp_python_stub",
-    SDK_DIR / "bondmcp-python.py",
-)
-bondmcp = importlib.util.module_from_spec(spec)
-spec.loader.exec_module(bondmcp)
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
+
+import bondmcp_sdk as bondmcp
 
 
 class DummyResponse:
@@ -42,7 +36,7 @@ def test_request_success(monkeypatch):
         calls["url"] = url
         return DummyResponse(200, {"ok": True})
 
-    monkeypatch.setattr(bondmcp.requests, "get", fake_get)
+    monkeypatch.setattr(bondmcp.client.requests, "get", fake_get)
     result = client.request("get", "/test")
     assert result == {"ok": True}
     assert calls["url"] == "https://example.com/test"
@@ -54,7 +48,7 @@ def test_request_http_error(monkeypatch):
     def fake_get(url, headers=None, params=None, timeout=None):
         return DummyResponse(404, {"error": {"message": "bad", "code": "404"}})
 
-    monkeypatch.setattr(bondmcp.requests, "get", fake_get)
+    monkeypatch.setattr(bondmcp.client.requests, "get", fake_get)
     with pytest.raises(bondmcp.BondMCPAPIError) as exc:
         client.request("get", "/test")
     assert exc.value.status_code == 404
@@ -67,6 +61,6 @@ def test_request_network_error(monkeypatch):
     def fake_get(url, headers=None, params=None, timeout=None):
         raise requests.exceptions.ConnectionError
 
-    monkeypatch.setattr(bondmcp.requests, "get", fake_get)
+    monkeypatch.setattr(bondmcp.client.requests, "get", fake_get)
     with pytest.raises(bondmcp.BondMCPNetworkError):
         client.request("get", "/test")
